% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ceteris_paribus_oscillations.R
\name{plot.ceteris_paribus_oscillations}
\alias{plot.ceteris_paribus_oscillations}
\title{Plot Ceteris Paribus Oscillations}
\usage{
\method{plot}{ceteris_paribus_oscillations}(x, ...)
}
\arguments{
\item{x}{a ceteris paribus oscillation explainer produced with function `calculate_oscillations()`}

\item{...}{other explainers that shall be plotted together}
}
\value{
a ggplot2 object
}
\description{
Function 'plot.ceteris_paribus_oscillations' plots variable importance plots.
}
\examples{
library("DALEX")
 \dontrun{
library("randomForest")
set.seed(59)

apartments_rf_model <- randomForest(m2.price ~ construction.year + surface + floor +
      no.rooms + district, data = apartments)

explainer_rf <- explain(apartments_rf_model,
      data = apartmentsTest, y = apartmentsTest$m2.price)

apartment <- apartmentsTest[1:2,]

cp_rf <- ceteris_paribus(explainer_rf, apartment)
plot(cp_rf, color = "_ids_")

vips <- calculate_oscillations(cp_rf)
vips
plot(vips)
}
}
