% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coral_plots.R
\name{plot.coral_explainer}
\alias{plot.coral_explainer}
\title{Plots Coral Explanations}
\usage{
\method{plot}{coral_explainer}(x, ..., plot_residuals = TRUE)
}
\arguments{
\item{x}{a ceteris paribus explainer produced with the 'ceteris_paribus' function}

\item{...}{other explainers that shall be plotted together}

\item{plot_residuals}{if TRUE (default) then residuals are plotted as red/blue bars}
}
\value{
a ggplot2 object
}
\description{
Function 'plot.coral_explainer_explainer' plots Coral Plots for a single prediction / observation.
}
\examples{
library("DALEX")
library("randomForest")
set.seed(59)

apartments_rf_model <- randomForest(m2.price ~ construction.year + surface + floor +
      no.rooms + district, data = apartments)

explainer_rf <- explain(apartments_rf_model,
      data = apartmentsTest[,2:6], y = apartmentsTest$m2.price)

new_apartment <- apartmentsTest[1, ]
new_apartment

cr_rf <- coral_plots(explainer_rf, observation = new_apartment, select_points = 0.002)
plot(cr_rf, plot_residuals = FALSE)
plot(cr_rf)

new_apartment <- apartmentsTest[10, ]
cr_rf <- coral_plots(explainer_rf, observation = new_apartment, select_points = 0.002)
plot(cr_rf, plot_residuals = FALSE)
plot(cr_rf)

new_apartment <- apartmentsTest[302, ]
cr_rf <- coral_plots(explainer_rf, observation = new_apartment, select_points = 0.002)
plot(cr_rf, plot_residuals = FALSE)
plot(cr_rf)

new_apartment <- apartmentsTest[720, ]
cr_rf <- coral_plots(explainer_rf, observation = new_apartment, select_points = 0.002)
plot(cr_rf, plot_residuals = FALSE)
plot(cr_rf)
}
