% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_what_if.R
\name{plot.what_if_explainer}
\alias{plot.what_if_explainer}
\title{Plots What-If Explanations}
\usage{
\method{plot}{what_if_explainer}(x, ...)
}
\arguments{
\item{x}{a what-if explainer produced with the 'what_if' function}

\item{...}{other explainers that shall be plotted together}
}
\value{
a ggplot2 object
}
\description{
Function 'plot.what_if_explainer' plots What-If plots for a single prediction.
}
