% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactive_what_if.R
\name{plot_interactive}
\alias{plot_interactive}
\alias{plot_interactive.what_if_explainer}
\alias{plot_interactive.default}
\title{Plots Interactive What-If Explanations}
\usage{
\method{plot_interactive}{what_if_explainer}(x, ..., split = "models")

plot_interactive(x, ...)

\method{plot_interactive}{default}(x, ..., split = "models")
}
\arguments{
\item{x}{a what-if explainer produced with the 'what_if' function}

\item{...}{other explainers that shall be plotted together}

\item{split}{a character, either 'models' or 'variables'. Sets the vaiable for faceting}
}
\value{
a ggiraph object
}
\description{
Function 'plot.what_if_explainer' plots What-If plots for a single prediction.
}
