% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactive_d3.R
\name{plot_interactive_d3}
\alias{plot_interactive_d3}
\alias{plot_interactive_d3.ceteris_paribus_explainer}
\alias{plot_interactive_d3.default}
\title{Plots D3-based Interactive Ceteris Paribus Explanations}
\usage{
\method{plot_interactive_d3}{ceteris_paribus_explainer}(model, ...,
  selected_variables = NULL, color = NULL, width = NULL,
  height = NULL, no_colors = NULL, categorical_order = NULL,
  size_rugs = NULL, alpha_rugs = NULL, color_rugs = NULL,
  color_points = NULL, color_residuals = NULL, color_pdps = NULL,
  alpha_residuals = NULL, alpha_points = NULL, alpha_ices = NULL,
  alpha_pdps = NULL, size_points = NULL, size_residuals = NULL,
  size_ices = NULL, size_pdps = NULL, show_profiles = TRUE,
  show_observations = TRUE, show_rugs = NULL, show_residuals = NULL,
  aggregate_profiles = NULL)

plot_interactive_d3(x, ...)

\method{plot_interactive_d3}{default}(model, ...,
  selected_variables = NULL, color = NULL, width = NULL,
  height = NULL, no_colors = NULL, categorical_order = NULL,
  size_rugs = NULL, alpha_rugs = NULL, color_rugs = NULL,
  color_points = NULL, color_residuals = NULL, color_pdps = NULL,
  alpha_residuals = NULL, alpha_points = NULL, alpha_ices = NULL,
  alpha_pdps = NULL, size_points = NULL, size_residuals = NULL,
  size_ices = NULL, size_pdps = NULL, show_profiles = TRUE,
  show_observations = TRUE, show_rugs = NULL, show_residuals = NULL,
  aggregate_profiles = NULL)
}
\arguments{
\item{model}{a ceteris paribus explainer produced with function `ceteris_paribus()`}

\item{...}{other explainers that shall be plotted together}

\item{selected_variables}{if not NULL then only `selected_variables` will be presented}

\item{color}{a character. Either name of a color or name of a variable that should be used for coloring}

\item{width}{a numeric. Width (in px) of the whole plot}

\item{height}{a numeric. Heigth (in px) of the whole plot}

\item{no_colors}{a numeric. Number of colors in legend for sequential scales}

\item{categorical_order}{a list. List with order of values for categorical variables}

\item{size_rugs}{a numeric. Size of rugs to be plotted}

\item{alpha_rugs}{a numeric between 0 and 1. Opacity of rug lines}

\item{color_rugs}{a character. Name of a color. If NULL elements are plotted according to 'color' arguments}

\item{color_points}{a character. Name of a color. If NULL elements are plotted according to 'color' arguments}

\item{color_residuals}{a character. Name of a color. If NULL elements are plotted according to 'color' arguments}

\item{color_pdps}{a character. Name of a color. If NULL elements are plotted according to 'color' arguments}

\item{alpha_residuals}{a numeric between 0 and 1. Opacity of residuals}

\item{alpha_points}{a numeric between 0 and 1. Opacity of points}

\item{alpha_ices}{a numeric between 0 and 1. Opacity of ICE lines}

\item{alpha_pdps}{a numeric between 0 and 1. Opacity of PDP lines}

\item{size_points}{a numeric. Size of points to be plotted}

\item{size_residuals}{a numeric. Size of residuals (lines and points) to be plotted}

\item{size_ices}{a numeric. Size of ICE lines to be plotted}

\item{size_pdps}{a numeric. Size of PDP lines to be plotted}

\item{show_profiles}{a logical. If TRUE then individual profiles will be plotted}

\item{show_observations}{a logical. If TRUE then individual observations will be marked as points}

\item{show_rugs}{a logical. If TRUE then individual observations will be marked as rugs}

\item{show_residuals}{a logical. If TRUE then residuals will be plotted as a line ended with a point}

\item{aggregate_profiles}{a character. Either 'mean' or 'median'. If not NULL then profiles will be aggregated according to chosen metric and the aggregate profiles will be plotted}
}
\value{
a ceterisParibusD3 object
}
\description{
Function 'plot_interactive_d3.ceteris_paribus_explainer' plots interactive version of Ceteris Paribus Plots for selected observations.
Various parameters help to decide what should be plotted, profiles, aggregated profiles, points or rugs.
}
\examples{
library("DALEX")
 \dontrun{
library("ceterisParibusD3")
library("randomForest")
set.seed(59)

apartments_rf_model <- randomForest(m2.price ~ construction.year + surface + floor +
                                     no.rooms + district,
                                   data = apartments)

explainer_rf <- explain(apartments_rf_model,
                       data = apartmentsTest[,2:6],
                       y = apartmentsTest$m2.price)

apartments_A <- apartmentsTest[958,]

cp_rf_A <- ceteris_paribus(explainer_rf, apartments_A, y = apartments_A$m2.price)

plot_interactive_d3(cp_rf_A, show_profiles = TRUE, show_observations = TRUE,
                selected_variables = c("surface","construction.year"))
}
}
