% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ceteris_paribus.R
\name{+.plot_ceteris_paribus_explainer}
\alias{+.plot_ceteris_paribus_explainer}
\title{Add More Layers to a Ceteris Paribus Plot}
\usage{
\method{+}{plot_ceteris_paribus_explainer}(e1, e2)
}
\arguments{
\item{e1}{An object of class `plot_ceteris_paribus_explainer`.}

\item{e2}{A plot component}
}
\description{
Add More Layers to a Ceteris Paribus Plot
}
