% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_what_if.R
\name{print.ceteris_paribus_explainer}
\alias{print.ceteris_paribus_explainer}
\title{Prints Ceteris Paribus Explainer Summary}
\usage{
\method{print}{ceteris_paribus_explainer}(x, ...)
}
\arguments{
\item{x}{a ceteris_paribus explainer produced with the 'ceteris_paribus' function}

\item{...}{other arguments that will be passed to head()}
}
\description{
Prints Ceteris Paribus Explainer Summary
}
\examples{
library("DALEX")
library("randomForest")
apartments_rf_model <- randomForest(m2.price ~ construction.year + surface + floor +
                     no.rooms + district, data = apartments)
explainer_rf <- explain(apartments_rf_model,
                     data = apartmentsTest[,2:6], y = apartmentsTest$m2.price)
new_apartment <- apartmentsTest[1, ]
new_apartment
}
