% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_coral_plots.R
\name{print.coral_plot_explainer}
\alias{print.coral_plot_explainer}
\title{Prints Coral Plots Explainer Summary}
\usage{
\method{print}{coral_plot_explainer}(x, ...)
}
\arguments{
\item{x}{a coral_plot explainer produced with the 'coral_plot' function}

\item{...}{other arguments that will be passed to head()}
}
\description{
Prints Coral Plots Explainer Summary
}
\examples{
library("DALEX")
library("randomForest")
apartments_rf_model <- randomForest(m2.price ~ construction.year + surface + floor +
                     no.rooms + district, data = apartments)
explainer_rf <- explain(apartments_rf_model,
                     data = apartmentsTest[,2:6], y = apartmentsTest$m2.price)
new_apartment <- apartmentsTest[1, ]
new_apartment
}
