% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ceteris_paribus.R
\name{print.plot_ceteris_paribus_explainer}
\alias{print.plot_ceteris_paribus_explainer}
\title{Print Ceteris Paribus Explainer Summary}
\usage{
\method{print}{plot_ceteris_paribus_explainer}(x, ...)
}
\arguments{
\item{x}{a plot_ceteris_paribus_explainer object to plot}

\item{...}{other arguments that will be passed to `print.ggplot()`}
}
\description{
See more examples in the \code{ceteris_paribus_layer} function
}
