% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_what_if.R
\name{print.what_if_explainer}
\alias{print.what_if_explainer}
\title{Prints What-If Explainer Summary}
\usage{
\method{print}{what_if_explainer}(x, ...)
}
\arguments{
\item{x}{a what-if explainer produced with the 'what_if' function}
}
\description{
Prints What-If Explainer Summary
}
