% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_if.R
\name{what_if}
\alias{what_if}
\title{What-If 1D Plot}
\usage{
what_if(explainer, observation, grid_points = 101)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the 'DALEX::explain' function}

\item{observation}{a new observarvation for which predictions need to be explained}

\item{grid_points}{number of points used for response path}
}
\value{
An object of the class 'what_if_explainer'.
It's a data frame with calculated average responses.
}
\description{
What-If 1D Plot
}
