% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_what_if.R
\name{print.what_if_explainer}
\alias{print.what_if_explainer}
\title{Print What If Explainer Summary}
\usage{
\method{print}{what_if_explainer}(x, ...)
}
\arguments{
\item{x}{a what_if explainer produced with the 'what_if' function}

\item{...}{other arguments that will be passed to head()}
}
\description{
Print What If Explainer Summary
}
\examples{
library("DALEX")
 \dontrun{
library("randomForest")
apartments_rf_model <- randomForest(m2.price ~ construction.year + surface + floor +
                     no.rooms + district, data = apartments)
explainer_rf <- explain(apartments_rf_model,
                     data = apartmentsTest[,2:6], y = apartmentsTest$m2.price)
new_apartment <- apartmentsTest[1, ]
new_apartment
}
}
